프로필에 내가 좋아요한 게시글 수/댓글 수 응답필드 추가하기

+++ src/main/java/com/sparta/ottoon/comment/dto/CommentResponseDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 public class CommentResponseDto {
     private Long commentId;
     private String comment;
-    private Long likeCount;
+    private int likeCount;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
