+++ src/main/java/com/sparta/ottoon/post/dto/PostResponseDto.java
@@ -20,7 +20,7 @@ public class PostResponseDto {
     private Long postId;
     @NotBlank(message = "내용을 입력해 주세요")
     private String contents;
-    private Long likeCount;
+    private int likeCount;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createdAt;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
